Move dir and cpacs selection on the main page
